+++ py-polars/src/lazyframe/mod.rs
@@ -2,9 +2,9 @@ mod exitable;
 mod visit;
 pub(crate) mod visitor;
 use std::collections::HashMap;
-use std::io::{BufReader, BufWriter};
 use std::num::NonZeroUsize;
 use std::path::PathBuf;
+mod serde;
 
 pub use exitable::PyInProcessQuery;
 use polars::io::cloud::CloudOptions;
@@ -15,13 +15,12 @@ use polars_core::prelude::*;
 use polars_parquet::arrow::write::StatisticsOptions;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
-use pyo3::pybacked::{PyBackedBytes, PyBackedStr};
-use pyo3::types::{PyBytes, PyDict, PyList};
+use pyo3::pybacked::PyBackedStr;
+use pyo3::types::{PyDict, PyList};
 pub(crate) use visit::PyExprIR;
 
 use crate::error::PyPolarsErr;
 use crate::expr::ToExprs;
-use crate::file::get_file_like;
 use crate::interop::arrow::to_rust::pyarrow_schema_to_rust;
 use crate::lazyframe::visit::NodeTraverser;
 use crate::prelude::*;
@@ -43,79 +42,6 @@ impl From<LazyFrame> for PyLazyFrame {
 #[pymethods]
 #[allow(clippy::should_implement_trait)]
 impl PyLazyFrame {
-    fn __getstate__(&self, py: Python) -> PyResult<PyObject> {
-        // Used in pickle/pickling
-        let mut writer: Vec<u8> = vec![];
-        ciborium::ser::into_writer(&self.ldf.logical_plan, &mut writer)
-            .map_err(|e| PyPolarsErr::Other(format!("{}", e)))?;
-
-        Ok(PyBytes::new_bound(py, &writer).to_object(py))
-    }
-
-    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
-        // Used in pickle/pickling
-        match state.extract::<PyBackedBytes>(py) {
-            Ok(s) => {
-                let lp: DslPlan = ciborium::de::from_reader(&*s)
-                    .map_err(|e| PyPolarsErr::Other(format!("{}", e)))?;
-                self.ldf = LazyFrame::from(lp);
-                Ok(())
-            },
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Serialize into binary data.
-    fn serialize_binary(&self, py_f: PyObject) -> PyResult<()> {
-        let file = BufWriter::new(get_file_like(py_f, true)?);
-        ciborium::into_writer(&self.ldf.logical_plan, file)
-            .map_err(|err| PyValueError::new_err(format!("{err:?}")))?;
-        Ok(())
-    }
-
-    /// Serialize into a JSON string.
-    #[cfg(feature = "json")]
-    fn serialize_json(&self, py_f: PyObject) -> PyResult<()> {
-        let file = BufWriter::new(get_file_like(py_f, true)?);
-        serde_json::to_writer(file, &self.ldf.logical_plan)
-            .map_err(|err| PyValueError::new_err(format!("{err:?}")))?;
-        Ok(())
-    }
-
-    /// Deserialize a file-like object containing binary data into a LazyFrame.
-    #[staticmethod]
-    fn deserialize_binary(py_f: PyObject) -> PyResult<Self> {
-        let file = get_file_like(py_f, false)?;
-        let reader = BufReader::new(file);
-        let lp = ciborium::from_reader::<DslPlan, _>(reader)
-            .map_err(|err| PyValueError::new_err(format!("{err:?}")))?;
-        Ok(LazyFrame::from(lp).into())
-    }
-
-    /// Deserialize a file-like object containing JSON string data into a LazyFrame.
-    #[staticmethod]
-    #[cfg(feature = "json")]
-    fn deserialize_json(py_f: PyObject) -> PyResult<Self> {
-        // it is faster to first read to memory and then parse: https://github.com/serde-rs/json/issues/160
-        // so don't bother with files.
-        let mut json = String::new();
-        let _ = get_file_like(py_f, false)?
-            .read_to_string(&mut json)
-            .unwrap();
-
-        // SAFETY:
-        // We skipped the serializing/deserializing of the static in lifetime in `DataType`
-        // so we actually don't have a lifetime at all when serializing.
-
-        // &str still has a lifetime. But it's ok, because we drop it immediately
-        // in this scope.
-        let json = unsafe { std::mem::transmute::<&'_ str, &'static str>(json.as_str()) };
-
-        let lp = serde_json::from_str::<DslPlan>(json)
-            .map_err(|err| PyValueError::new_err(format!("{err:?}")))?;
-        Ok(LazyFrame::from(lp).into())
-    }
-
     #[staticmethod]
     #[cfg(feature = "json")]
     #[allow(clippy::too_many_arguments)]
